Parameterize the max line length for the report table.
This adds an extra iteration across all violations in all cases. It’s 
possible to only do so when the max line length is specified, but doing 
so increases the number of branches and makes the code a bit harder to
read.

This also creates the FlipTable twice, once to measure the line lengths 
and then again after we’ve set our target line breaks. It’s also 
possible to make some assumptions about the whitespace and number of 
characters that FlipTable will add and avoid the second table creation. 
However, assuming that the construction of the table is not super 
performance critical, it seems more robust to just measure the output.

A nice TODO might be to add coherent line breaking so we don’t break
in the middle of words, but this isn’t technically a regression from
the current 100 character message limit.