The CppCheck parser should retain the order of violations
The CppCheck parser handels individual locations of a warning as separate violations.
These violations are then returned to the caller as a sorted set, destroying the insertion order.
To visualize the affected locations in the original order, we need to store the insertion order
by adding a new property.

See https://issues.jenkins.io/browse/JENKINS-75217